backend/requirements.txt
# Telegram Bot
aiogram==3.4.1

# Web Framework
fastapi==0.109.2
uvicorn==0.27.1

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Scheduler
apscheduler==3.10.4

# Utils
python-dateutil==2.8.2
pytz==2024.1
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0